extract io
